requirements.txt
pyjce2
pytea2
pydantic
loguru
pydispatcher
aiohttp